requirements.txt
flask==2.2.*
gunicorn==20.1.*
isodate==0.6.*
pyemvue==0.15.*
python-decouple==3.6.*
pytz==2022.*
